WebSockets/tests/submissions/Microsoft/Secure-Close-Reason-Unpaired-surrogates.htm

<!DOCTYPE html>
<html>
<head>
    <title>W3C WebSocket API - Close Secure WebSocket - Reason with unpaired surrogates</title>
    <script type="text/javascript" src="http://w3c-test.org/resources/testharness.js"></script>
    <script type="text/javascript" src="http://w3c-test.org/resources/testharnessreport.js"></script>
    <script type="text/javascript" src="websocket.js"></script>
</head>
<body>
    <div id="log"></div>
    <script type="text/javascript">

        var test = async_test("W3C WebSocket API - Create Secure WebSocket - Close the Connection - close(code, 'reason with unpaired surrogates') - SYNTAX_ERR is thrown");

        var wsocket = CreateWebSocket(true, false, false);
        var isOpenCalled = false;

        wsocket.addEventListener('open', test.step_func(function (evt) {
            var reason = "\uD807";
            assert_throws("SYNTAX_ERR", function () { wsocket.close(1000, reason) });
            clearTimeout(timeOut);
            test.done();
        }), true);
    </script>

</body>
</html>
